requirements.txt
autopep8==1.5.6
azure-core==1.12.0
azure-cosmos==4.2.0
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
fastapi==0.63.0
Flask==1.1.2
gunicorn==20.0.4
h11==0.12.0
httptools==0.1.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycodestyle==2.7.0
pydantic==1.8.1
redis==3.5.3
requests==2.25.1
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
Werkzeug==1.0.1
